// packages/detail/src/types/movie.ts
export interface Movie {
  id: number
  title: string
  poster_path: string
  overview: string
  release_date: string
  vote_average: number
  genres?: { id: number; name: string }[]
  runtime?: number
  budget?: number
  revenue?: number
  status?: string
  tagline?: string
}

export type MovieDetailResponse = Movie